gitignoreに不要なファイルを追記しました。 キャラクターがめり込まないよう、壁の位置を微調整しました。 モーションファイルから攻撃パーツの番号を取得しました。

+++ source/debugProc.cpp
@@ -86,8 +86,8 @@ void CDebugProc::Print(char * fmt, ...)
 	va_list list;					// ’è‹`‚·‚é‚¾‚¯‚Å’l‚ÍÝ’è‚µ‚È‚¢
 	int nCount;
 	char * cp;
-	int nLengh = strlen(m_aStr);	// •¶Žš‚ÌŒÂ”‚ð”‚¦‚é
-	int nFmtLengh = strlen(fmt);	
+	int nLengh = (int)strlen(m_aStr);	// •¶Žš‚ÌŒÂ”‚ð”‚¦‚é
+	int nFmtLengh = (int)strlen(fmt);	
 	if ((nLengh + nFmtLengh) > MAX_CHARACTER)
 	{
 		cout << "’´‚¦‚½\n";
